Relocate rule-tests under OpeningPeriodForm.test
